Fix placement added date cleared after bulk update
When bulk update of placements is performed the addedDate field of the
updated placements is deleted.

Update pom.xml to bump the version of tcs-client from 4.0.1 to 4.0.2.

Update UploadFileResourceTest and UploadFileServiceImplTest to change
the way the upload files are loaded to avoid a test failure when running
on a windows environment.