Refactor column mappers to use mapping data object
The column mappers currently use two string arrays to define column
mappings between source data and local variables. This is prone to
errors and makes it more difficult than it should be to quickly check
mappings. Improvement can be achieved by using a data object to wrap
matching entries from both arrays in a single object.

Create ColumnMapping as a data object containing targetFieldName,
sourceFieldName and a flag for whether the field is required.

Update the existing column mappers to use the ColumnMapping data object
instead of multiple arrays.

Update ColumnMapper to use the ColumnMapping list instead of the arrays.