Hook up Post Funding Update to TCS endpoint
Update PostFundingUpdateTransformerService to call the TCS endpoint for
patching post fundings. Validation should be done first to ensure that
the trust name can be converted to a trust ID.

Update the service pom.xml to bump the reference-client version to allow
trusts to be bulk searched by trustKnownAs.

Update TemplateXLS to add a method for adding error messages to simplify
the process.