feat: Add Error if row has incorrect role
If a row (Person) has incorrect role (i.e.: anything that doesn't
look like "DR in Training") should get an error added.
If a row (Person) has something that looks like "DR in Training"
but with inaccurate casing or stray spaces and dots, it gets fixed
on the go and no error is added.

TIS21-1629: DR in Training validation