test: validate converting Long types
This sets out what values are acceptable to match on.
Some code choices are made to boost coverage.

refactor: tweak Mapper "empty checks"

I do assume that the existing empty checks assume a string argument.
There is a subtle semantic difference between that and `toString()`.
With a lack of tests this is a more cautious change.

TIS21-7129: Use Person ID to create a placement against