fix(test): Clean up misleading Type parameters
Type parameters added were not appropriate for expected invocations

chore(version): major change to project POM

The removal of the client module is, theoretically, a breaking change.
The artifact is not a dependency in SCM.
Even if it were used, it would only specify transitive dependencies.

TIS21-7129: Use Person ID to create a placement against