fix: use compatible profile client
This quickly restores service for the stage/test environment.
The `profile-client` v3.2+ is built on Java 11.
We will be migrating generic upload to a later version of java.
The runtime JRE is derived from a custom base image.
There is an unknown amount of work to upgrade the docker image.

TIS21-7129: Use Person ID to create a placement against